+++ R/or_plot.R
@@ -77,6 +77,10 @@ or_plot = function(.data, dependent, explanatory, random_effect=NULL,
 		factorlist = summary_factorlist(.data, dependent, explanatory, total_col=TRUE, fit_id=TRUE)
 	}
 	
+	# For continuous variables, remove level label
+	drop = grepl("Mean \\(SD\\)|Median \\(IQR\\)", factorlist$levels)
+	factorlist$levels[drop] = "-"
+	
 	if(remove_ref){
 		factorlist = factorlist %>%  
 			dplyr::mutate(label = ifelse(label == "", NA, label)) %>%